Middleware config (#288)
* fix: middleware config

* fix: middleware config

* fix: rollback to 2.5.6

* fix(cms): deployment issue with cms
